rag-iep-system/services/doc-parser/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.1
langchain==0.0.335
langdetect==1.0.9
PyPDF2==3.0.1
python-docx==1.0.1
structlog==23.2.0
